# openapi.yaml
openapi: 3.0.3
info:
  title: SOC Application API
  description: |
    Security Operations Center (SOC) application for analyzing Zscaler NSS web proxy logs.
    
    This API provides endpoints for:
    - User authentication
    - Log file upload and management
    - Log entry analysis and filtering
    - Dashboard statistics and analytics
    - Anomaly detection
    - AI-powered threat analysis
    
    ## Authentication
    Most endpoints require JWT authentication. Obtain a token by calling the `/api/auth/login` endpoint,
    then include it in the `Authorization` header as `Bearer <token>`.
    
  version: 1.0.0
  contact:
    name: SOC Application Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/anomalies:
    get:
      tags:
        - Anomalies
      summary: List anomalies
      description: Get a paginated list of detected anomalies with optional filters
      operationId: listAnomalies
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: anomaly_type
          in: query
          description: Filter by anomaly type
          schema:
            type: string
            enum: [burst_blocked, malicious_domain, risky_category, unusual_ua, large_download]
        - name: min_confidence
          in: query
          description: Minimum confidence threshold (0.0-1.0)
          schema:
            type: number
            format: float
            default: 0.5
            minimum: 0.0
            maximum: 1.0
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of anomalies
          content:
            application/json:
              schema:
                type: object
                properties:
                  anomalies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Anomaly'
                  total:
                    type: integer
                    example: 25
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/anomalies/timeline:
    get:
      tags:
        - Anomalies
      summary: Get anomaly timeline
      description: Get timeline data showing anomalies over time
      operationId: getAnomalyTimeline
      parameters:
        - name: log_file_id
          in: query
          description: Filter by log file ID
          schema:
            type: string
            format: uuid
        - name: bucket_minutes
          in: query
          description: Time bucket size in minutes
          schema:
            type: integer
            default: 15
            minimum: 1
      responses:
        '200':
          description: Anomaly timeline data
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnomalyTimelineBucket'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
...
  schemas:
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        log_file_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        location:
          type: string
          example: ny-gre
        protocol:
          type: string
          example: HTTPS
        url:
          type: string
          example: https://example.com/path
        domain:
          type: string
          example: example.com
        action:
          type: string
          enum: [Allowed, Blocked]
        app_name:
          type: string
          example: Gmail
        app_class:
          type: string
          example: Business Apps
        throttle_req_size:
          type: integer
        throttle_resp_size:
          type: integer
        req_size:
          type: integer
        resp_size:
          type: integer
        url_class:
          type: string
          example: Business Use
        url_supercat:
          type: string
          example: Business and Economy
        url_cat:
          type: string
          example: Email
        dlp_dict:
          type: string
        dlp_eng:
          type: string
        dlp_hits:
          type: integer
        file_class:
          type: string
        file_type:
          type: string
        location2:
          type: string
        department:
          type: string
          example: Engineering
        client_ip:
          type: string
          format: ipv4
          example: 172.17.3.49
        server_ip:
          type: string
          format: ipv4
          example: 66.211.175.229
        http_method:
          type: string
          example: GET
        http_status:
          type: integer
          example: 200
        user_agent:
          type: string
          example: Mozilla/5.0
        threat_category:
          type: string
          example: Phishing
        fw_filter:
          type: string
        fw_rule:
          type: string
        policy_type:
          type: string
        reason:
          type: string
        is_anomalous:
          type: boolean
        anomaly_type:
          type: string
          enum: [burst_blocked, malicious_domain, risky_category, unusual_ua, large_download]
        anomaly_reason:
          type: string
        anomaly_confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        created_at:
          type: string
          format: date-time
...
    Anomaly:
      type: object
      properties:
        entry_id:
          type: string
          format: uuid
        log_entry:
          $ref: '#/components/schemas/LogEntry'
        anomaly_type:
          type: string
          enum: [burst_blocked, malicious_domain, risky_category, unusual_ua, large_download]
        reason:
          type: string
          example: Multiple blocked requests in short time window
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.85
    AnomalyTimelineBucket:
      type: object
      properties:
        time:
          type: string
          format: date-time
          example: "2022-06-20T12:00:00Z"
        count:
          type: integer
          example: 5
        by_type:
          type: object
          additionalProperties:
            type: integer
          example:
            malicious_domain: 3
            burst_blocked: 2
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...